added in old media tracker
