+++ talos.py
@@ -1,4 +1,5 @@
 from tracking.haar_cascade.basic_tracker import *
+from tracking.media_pipe.media_pipe_pose import *
 from tracking.media_pipe.media_pipe_tracker import *
 from tracking.yolo.yolo_tracker import *
 from directors.continuous_director import *
@@ -15,7 +16,7 @@ def main():
     #interface = ManualInterface()
     #interface.launch_user_interface()
 
-    
+    #tracker = MediaPipePose(args.source, "./config.yaml")
     tracker = MediaPipeTracker(args.source, "./config.yaml")
     #tracker = BasicTracker(args.source, "./config.yaml")
     #tracker = YOLOTracker(args.source, "./config.yaml")
